Fixing dashboard.html
